+++ Core/CrossCuttingConcerns/Caching/ICacheManager.cs
@@ -6,7 +6,7 @@ namespace Core.CrossCuttingConcerns.Caching
 {
     public interface ICacheManager
     {
-        void Add(string key, object value, int duration);
+        void Add(string key, object value, double duration);
         bool IsAdd(string key);
         T Get<T>(string key);
         object Get(string key);
